+++ tests/cli/sqls/select.sql
@@ -1,1 +1,2 @@
-select name, age from cli_test
+select name, age from cli_test;
+-- the semicolon and this comment are intentional :)